1. format code 2. add api for create new expense

+++ db_scripts/001_create_table_users.sql
@@ -37,8 +37,8 @@ create table expenses
     account_id  integer not null,
     cat_id      integer not null,
     subcat_id   integer null,
-    amount      integer null,
+    amount      REAL null,
     currency_id integer null,
     note        text null
 );
-create index exp_idx_user_id on expenses(user_id);
+create index exp_idx_user_id on expenses (user_id);
